+++ src/emitter.js
@@ -33,7 +33,7 @@ var emitterProto = {
 			;
 
 			for (i = listeners.length - 1; i >= 0; i--) {
-				if( listeners[i] === listener )
+				if( listeners[i].callback === listener )
 					listeners.splice( i, 1 );
 			}
 		}